Move feed discovery backends to their own module
