+++ net/discovery/README.md
@@ -1,0 +1,3 @@
+# :net:discovery
+
+This module contains the podcast search/discovery APIs.
