+++ net/discovery/src/main/java/de/danoeh/antennapod/net/discovery/PodcastSearcher.java
@@ -1,4 +1,4 @@
-package de.danoeh.antennapod.discovery;
+package de.danoeh.antennapod.net.discovery;
 
 import io.reactivex.Single;
 import java.util.List;
